consumer/requirements.txt
kafka-python
SQLAlchemy
psycopg2-binary
fastavro